fix: Using the correct click event (btnClick) in all templates
This commit fixes an issue where (click) events in the app-button component were not fired correctly.
The cause of the problem was the use of (click) event. This caused unexpected behavior by using event.stopPropagation() in the button element, which prevented the event from propagating to the app-button component.

SCHUL-27