update: Make holiday.component.spec.ts test runnable
This commit updates the 'click upload button and check if dialog exist' test in the HolidayComponent. The test now properly simulates a click event using the onClick method of the ButtonComponent, and checks if the btnClick output has been emitted. It also checks if a MatDialog instance is opened as expected. This ensures that the functionality works as intended across different conditions.

SCHUL-27

Co-Authored-By: Philipp Wildzeiss <54113330+phwildzeiss@users.noreply.github.com>