add: Router link directive stub
This commit adds the RouterLinkDirectiveStub class. It is a directive stub for the routerLink directive in Angular. The stub provides an input property linkParams to bind the router link parameters and a property navigatedTo to store the navigated link value. The onClick method is triggered when the host element is clicked, assigning the value of linkParams to navigatedTo. This stub is useful for testing components that utilize the routerLink directive.

SCHUL-160